requirements.txt
anyio==3.6.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
fastapi==0.87.0
graphql-core==3.2.3
h11==0.14.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
strawberry-graphql==0.142.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
websockets==10.4
